+++ fastapi_app/src/api/locations.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 from typing import List
 from src.schemas.location import Location, LocationCreate, LocationUpdate
 from src.repositories.location_repository import LocationRepository
@@ -13,6 +13,8 @@ from src.exceptions import (
     NotFoundError, UniqueConstraintError, ValidationError,
     NotFoundHTTPError, ConflictHTTPError, BadRequestHTTPError
 )
+from src.auth.dependencies import get_current_user
+from src.schemas.users import User 
 
 
 router = APIRouter(prefix="/locations", tags=["Locations"])
@@ -45,8 +47,10 @@ def get_location_by_id(location_id: int):
 
 
 @router.post("/", response_model=Location, status_code=status.HTTP_201_CREATED)
-def create_location(location_data: LocationCreate):
+def create_location(location_data: LocationCreate,
+                    current_user: User = Depends(get_current_user)):
     """Создать новую локацию"""
+
     repository = get_location_repository()
     use_case = CreateLocationUseCase(repository)
     try:
@@ -58,8 +62,11 @@ def create_location(location_data: LocationCreate):
 
 
 @router.put("/{location_id}", response_model=Location)
-def update_location(location_id: int, location_data: LocationUpdate):
+def update_location(location_id: int, location_data: LocationUpdate,
+                    current_user: User = Depends(get_current_user)):
     """Обновить локацию"""
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=403, detail="Admin access required")
     repository = get_location_repository()
     use_case = UpdateLocationUseCase(repository)
     
@@ -72,8 +79,11 @@ def update_location(location_id: int, location_data: LocationUpdate):
 
 
 @router.delete("/{location_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_location(location_id: int):
+def delete_location(location_id: int,
+                    current_user: User = Depends(get_current_user)):
     """Удалить локацию"""
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=403, detail="Admin access required")
     repository = get_location_repository()
     use_case = DeleteLocationUseCase(repository)
     try:
